+++ app/module/vike/vike.type.ts
@@ -2,25 +2,16 @@ import type { ComponentPublicInstance } from 'vue';
 
 type Component = ComponentPublicInstance;
 export type Page = Component;
-// eslint-disable-next-line unicorn/prevent-abbreviations
 type PageProps = NonNullable<unknown>;
 
 // https://vike.dev/pageContext#typescript
 
 declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
   namespace Vike {
     interface PageContext {
-      // @ts-ignore
       Page: Page;
       pageProps?: PageProps;
       urlPathname: string;
-      exports: {
-        documentProps?: {
-          title?: string;
-          description?: string;
-        };
-      };
     }
   }
 }