+++ renderer/onRenderHtml.ts
@@ -5,6 +5,7 @@ import type { PageContextServer } from 'vike/dist/esm/types';
 import { createApp } from 'vue';
 import { renderToString } from '../app/module/helper/helper.render.ts';
 import TEMPLATE_BASE from '../app/module/template/template.base';
+import { setPageContext } from '../app/module/vike/vike.vue-context.ts';
 
 interface PageCtx extends PageContextServer {
   pageProps?: NonNullable<unknown> | undefined;
@@ -25,9 +26,10 @@ async function onRenderHtml(
   } = pageContext;
   if (!Page) throw new Error('My render() hook expects pageContext.Page to be defined');
 
-  // @ts-ignore
-  // eslint-disable-next-line @typescript-eslint/no-unsafe-argument,@typescript-eslint/no-unsafe-call
-  const app = createApp(Page, pageProps, pageContext);
+  const app = createApp(Page, pageProps);
+  await (config as any).onCreateApp?.(app, pageContext);
+
+  await setPageContext(app, pageContext)
   const appHtml = Page ? await renderToString(app) : ''; // SSR | SPA
 
   // // See https://vike.dev/head