+++ tsup.config.ts
@@ -18,6 +18,7 @@ export default defineConfig({
   sourcemap: true,
   minify: true,
   clean: true,
+  external: ['vue', 'vike']
   // banner: {
   //   js: `import{fileURLToPath}from"node:url";import path from"node:path";import{createRequire as topLevelCreateRequire}from"module";const require=topLevelCreateRequire(import.meta.url),__filename=fileURLToPath(import.meta.url),__dirname=path.dirname(__filename);`,
   // },